+++ API/Escola.Service/Service/TurmaService.cs
@@ -7,9 +7,10 @@ using Escola.Domain.Interface.Repository;
 
 namespace Escola.Application.Service
 {
-    public class TurmaService(ITurmaRepository turmaRepository) : ITurmaService
+    public class TurmaService(ITurmaRepository turmaRepository, IAlunoTurmaRepository alunoTurmaRepository) : ITurmaService
     {
         private readonly ITurmaRepository _turmaRepository = turmaRepository;
+        private readonly IAlunoTurmaRepository _alunoTurmaRepository = alunoTurmaRepository;
         private readonly List<Erro> _erros = [];
 
         public async Task<TurmaResponse> AddAsync(TurmaRequest turmaRequest)
@@ -23,8 +24,8 @@ namespace Escola.Application.Service
             var turma = new Turma()
             {
                 Nome = turmaRequest.Nome,
-                CursoId = turmaRequest.CursoId,
-                Ano = turmaRequest.Ano
+                CursoId = turmaRequest.CursoId ?? 0,
+                Ano = turmaRequest.Ano ?? DateTime.Now.Year
             };
 
             turma = await _turmaRepository.AddAsync(turma);
@@ -35,6 +36,9 @@ namespace Escola.Application.Service
         public async Task DeleteByIdAsync(int id)
         {
             _ = await _turmaRepository.GetByIdAsync(id) ?? throw new NotFoundException(DomainErrors.Turma.NotFound);
+
+            var alunos = await _alunoTurmaRepository.GetByAlunoId(id);
+            if (alunos.Any()) throw new BadRequestException(DomainErrors.Turma.HasAluno);
 
             await _turmaRepository.DeleteByIdAsync(id);
         }
@@ -57,7 +61,12 @@ namespace Escola.Application.Service
 
         public async Task<TurmaResponse> UpdateAsync(TurmaRequest turmaRequest)
         {
-            var turma = await _turmaRepository.GetByIdAsync(turmaRequest.Id) ?? throw new NotFoundException(DomainErrors.Turma.NotFound);
+            var turma = await _turmaRepository.GetByIdAsync(turmaRequest.Id ?? 0) ?? throw new NotFoundException(DomainErrors.Turma.NotFound);
+
+            if (turmaRequest.Nome != turma.Nome && await _turmaRepository.NomeExistsAsync(turmaRequest.Nome)) _erros.Add(DomainErrors.Turma.NomeExists);
+            if (turmaRequest.Ano != turma.Ano && turmaRequest.Ano < DateTime.Now.Year) _erros.Add(DomainErrors.Turma.InvalidAno);
+
+            if (_erros.Count != 0) throw new BadRequestException(_erros);
 
             await _turmaRepository.UpdateAsync(turma);
